Climate model: Add ability to hide 90% of agents
We go through each agent set in turn and hide 90%,
rounded-down. This ensures that there is always at
least one agent of each breed visible.

If we are hiding, when new sunrays come in they have
a 90% change of being hidden.

New CO2 and Clouds come in visible.

Extra care is taken to ensure only whole clouds are
hidden, treating each cloud as an individual agent.