Draw initial scene before we start run() loop
Because of the timing differences between the run() and the 
animate() loops, @ticks was usually > 1 on the first call 
of draw(). This meant that models that do not automatically 
refresh patches would never draw their initial state.

Now we call draw() once before starting the model, to
paint the initial scene.