Extract generic ClimateModel to own script; extend it for Ocean Model
Basic Climate Model script is added as a new external script that is
loaded and parsed by coffeescript.

Because doing it this way appears to allow body.onload to be triggered
before script is loaded, we currently wait for script to be loaded by
using a $.Deferred pattern. It would be good to see if there's another
way we could accomplish this without needing this.

OceanClimateModel extends ClimateModel and adds only what is needed for
the ocean and water vapor. Again, we need to add a second $.Deferred.

NOTE: It may be necessary to add the MIME type .coffee => text/x-coffeescript
to your server's configuration. However, this should not be necessary.